+++ server/src/user/info.rs
@@ -18,12 +18,12 @@ pub async fn get_user_info(
 
     let user_id = user.id;
 
-    let user_info = sqlx::query!("SELECT * FROM users WHERE id = $1", user_id)
+    let user_info = sqlx::query!("SELECT * FROM users WHERE id = $1", user_id as i32)
         .fetch_one(db_pool)
         .await?;
 
     Ok(Json(UserInfo {
-        id: user_info.id,
+        id: user_info.id as u32,
         login: user_info.name,
         avatar_url: user_info.avatar_url,
         is_admin: user_info.id == 90502461,
